requirements.txt
langchain==1.2.1
openai==2.14.0
langchain-openai==1.1.6
google-api-python-client>=2.187.0
google-auth>=2.42.0,<3.0.0
google-auth-oauthlib>=1.2.0
python-dotenv==1.2.1
gradio==6.2.0
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
alembic==1.17.2
tzlocal==5.1
dateparser==1.1.8